Use only pandoc
asciidoctor-epub3 has some issues: the document must be splitted into
chapter files, cover images don't work, the fonts are embedded, the
default style is not simple, the stylesheets are too complex...

It seems a future version of asciidoctor will create EPUB directly.

At the moment, pandoc is much simpler and the resulting EPUB is good
enough.